+++ src/sdk/base/VerifyingPaymasterAPI.ts
@@ -1,5 +1,5 @@
-import axios from 'axios';
 import { ethers } from 'ethers';
+import fetch from 'cross-fetch';
 import { calcPreVerificationGas } from './calcPreVerificationGas';
 import { PaymasterAPI } from './PaymasterAPI';
 import { UserOperationStruct } from '../contracts/account-abstraction/contracts/core/BaseAccount';
@@ -10,20 +10,27 @@ const DUMMY_PAYMASTER_AND_DATA =
   '0x0101010101010101010101010101010101010101000000000000000000000000000000000000000000000000000001010101010100000000000000000000000000000000000000000000000000000000000000000101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101010101';
 
 export interface PaymasterResponse {
-  paymasterAndData: string;
-  verificationGasLimit: string;
-  preVerificationGas?: string;
+  result: {
+    paymasterAndData: string;
+    verificationGasLimit: string;
+    preVerificationGas: string;
+    callGasLimit: string;
+  }
 }
 
 export class VerifyingPaymasterAPI extends PaymasterAPI {
   private paymasterUrl: string;
   private entryPoint: string;
   private context: any;
-  constructor(paymasterUrl: string, entryPoint: string, context: any) {
+  private api_key: string;
+  private chainId: number;
+  constructor(paymasterUrl: string, entryPoint: string, context: any, api_key: string, chainId: number) {
     super();
     this.paymasterUrl = paymasterUrl;
     this.entryPoint = entryPoint;
     this.context = context;
+    this.api_key = api_key;
+    this.chainId = chainId;
   }
 
   async getPaymasterAndData(userOp: Partial<UserOperationStruct>): Promise<PaymasterResponse> {
@@ -32,7 +39,7 @@ export class VerifyingPaymasterAPI extends PaymasterAPI {
       // userOp.preVerificationGas contains a promise that will resolve to an error.
       await ethers.utils.resolveProperties(userOp);
       // eslint-disable-next-line no-empty
-    } catch (_) {}
+    } catch (_) { }
     const pmOp: Partial<UserOperationStruct> = {
       sender: userOp.sender,
       nonce: userOp.nonce,
@@ -50,20 +57,28 @@ export class VerifyingPaymasterAPI extends PaymasterAPI {
     op.preVerificationGas = calcPreVerificationGas(op);
 
     // Ask the paymaster to sign the transaction and return a valid paymasterAndData value.
-    const paymasterAndData = await axios
-      .post<PaymasterResponse>(this.paymasterUrl, {
-        jsonrpc: '2.0',
-        id: 1,
-        method: 'pm_sponsorUserOperation',
-        params: [await toJSON(op), this.entryPoint, this.context],
+    const paymasterAndData = await fetch(this.paymasterUrl, {
+      method: 'POST',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({ params: [await toJSON(op), this.entryPoint, this.context, this.chainId, this.api_key], jsonrpc: '2', id: 2 }),
+    })
+      .then(async (res) => {
+        const response = await await res.json();
+        if (response.error) {
+          throw new Error(response.error);
+        }
+        return response
       })
-      .then((res) => {
-        return res.data
-      });
+      .catch((err) => {
+        throw new Error(err.message);
+      })
 
-    return {paymasterAndData: paymasterAndData.paymasterAndData, verificationGasLimit: paymasterAndData.verificationGasLimit, preVerificationGas: op.preVerificationGas.toString()};
+    return paymasterAndData;
   }
 }
 
-export const getVerifyingPaymaster = (paymasterUrl: string, entryPoint: string, context: any) =>
-  new VerifyingPaymasterAPI(paymasterUrl, entryPoint, context);
+export const getVerifyingPaymaster = (paymasterUrl: string, entryPoint: string, context: any, api_key: string, chainId: number) =>
+  new VerifyingPaymasterAPI(paymasterUrl, entryPoint, context, api_key, chainId);