+++ src/sdk/network/constants.ts
@@ -75,7 +75,7 @@ export const Networks: {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       walletFactory: '0x7f6d8F107fE8551160BD5351d5F1514A6aD5d40E',
     },
-    graphqlEndpoint: '',
+    graphqlEndpoint: 'qa-etherspot.pillarproject.io',
   },
   [80001]: {
     chainId: 80001,
@@ -84,7 +84,7 @@ export const Networks: {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       walletFactory: '0x7f6d8F107fE8551160BD5351d5F1514A6aD5d40E',
     },
-    graphqlEndpoint: '',
+    graphqlEndpoint: 'qa-etherspot.pillarproject.io',
   },
   [84531]: {
     chainId: 84531,
@@ -93,7 +93,7 @@ export const Networks: {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       walletFactory: '0x7f6d8F107fE8551160BD5351d5F1514A6aD5d40E',
     },
-    graphqlEndpoint: '',
+    graphqlEndpoint: 'qa-etherspot.pillarproject.io',
   },
   [11155111]: {
     chainId: 11155111,
@@ -147,7 +147,7 @@ export const Networks: {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       walletFactory: '0x7f6d8F107fE8551160BD5351d5F1514A6aD5d40E',
     },
-    graphqlEndpoint: '',
+    graphqlEndpoint: 'qa-etherspot.pillarproject.io',
   },
   [10200]: {
     chainId: 10200,
@@ -156,7 +156,7 @@ export const Networks: {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       walletFactory: '0x7f6d8F107fE8551160BD5351d5F1514A6aD5d40E',
     },
-    graphqlEndpoint: '',
+    graphqlEndpoint: 'qa-etherspot.pillarproject.io',
   },
   [122]: {
     chainId: 122,
@@ -196,12 +196,12 @@ export const Networks: {
   },
   [420]: {
     chainId: 420,
-    bundler: '',
+    bundler: 'https://optimismgoerli-bundler.etherspot.io/',
     contracts: {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       walletFactory: '0x7f6d8F107fE8551160BD5351d5F1514A6aD5d40E',
     },
-    graphqlEndpoint: '',
+    graphqlEndpoint: 'qa-etherspot.pillarproject.io',
   },
   [31]: {
     chainId: 31,
@@ -273,7 +273,7 @@ export const Networks: {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       walletFactory: '0x7f6d8F107fE8551160BD5351d5F1514A6aD5d40E'
     },
-    graphqlEndpoint: ''
+    graphqlEndpoint: 'qa-etherspot.pillarproject.io'
   },
   [43113]: {
     chainId: 43113,
@@ -282,7 +282,7 @@ export const Networks: {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       walletFactory: '0x7f6d8F107fE8551160BD5351d5F1514A6aD5d40E'
     },
-    graphqlEndpoint: ''
+    graphqlEndpoint: 'qa-etherspot.pillarproject.io'
   },
   [59144]: {
     chainId: 59144,